demo4/requirements.txt
Django>=4.0
PyJWT>=2.0.0
